Tried a bunch of different Romanian dictionaries (ispell, huspell), but none were better than the vanilla aspell one. Sorry :/.
But it seems the Romanian ignore list is full of duplicates as well as words
that don't need to be ignored.  The current templates only contain ~350 or
so words not in the official aspell romanian dictionary, while there were
well over 2000 of them in the Romaninan ignore file.  Some were duplicates,
but not all (see diff).  So I shorthened the ignore file to only include
stuff that's currently not in the dictionary.

Not sure if this is what you want (the other ignore words might have been
needed before and might be needed again).

Ref: #40